refactor: Entity 기준으로 유틸 함수 분리
