+++ src/refactoring/hooks/utils/product/index.ts
@@ -1,0 +1,5 @@
+import { Product } from '../../../../types'
+
+export const findProductById = (products: Product[], productId: string): Product | undefined => {
+  return products.find((p) => p.id === productId)
+}
