+++ src/refactoring/hooks/utils/services/stockManagementService.ts
@@ -1,0 +1,7 @@
+import { CartItem, Product } from "../../../../types"
+
+export const getRemainingStock = (cart: CartItem[], product: Product) => {
+    const cartItem = cart.find((item) => item.product.id === product.id)
+    return product.stock - (cartItem?.quantity || 0)
+  }
+
